feat(group): 批量删除k8s/主机规则组，内置规则组不允许被删除

+++ hummer-modules/hummer-cloud/src/main/java/com/hummer/cloud/service/RuleService.java
@@ -728,9 +728,11 @@ public class RuleService {
 
     public void deleteRuleGroupById(Integer id, LoginUser loginUser) {
         RuleGroup ruleGroup = ruleGroupMapper.selectByPrimaryKey(id);
-        ruleGroupMapper.deleteByPrimaryKey(id);
-        operationLogService.log(loginUser, ruleGroup.getId().toString(), ruleGroup.getName(), ResourceTypeConstants.RULE_GROUP.name(), ResourceOperation.DELETE, "i18n_delete_delete_rule_group");
-
+        //内置规则组不可以被删除
+        if(!ruleGroup.getFlag()) {
+            ruleGroupMapper.deleteByPrimaryKey(id);
+            operationLogService.log(loginUser, ruleGroup.getId().toString(), ruleGroup.getName(), ResourceTypeConstants.RULE_GROUP.name(), ResourceOperation.DELETE, "i18n_delete_delete_rule_group");
+        }
     }
 
     public void deleteGroups(List<Integer> ids, LoginUser loginUser) throws Exception {
@@ -742,7 +744,7 @@ public class RuleService {
             }
         });
     }
-    
+
     public List<GroupDTO> groups(List<String> ids) {
         List<GroupDTO> groupDTOS = new LinkedList<>();
         for (String id : ids) {
